refactor: second test
